pyproject.toml
[project]
authors = [{ name = "Joseph Ryan-Palmer", email = "joseph@ryan-palmer.com" }]
version = "0.1.4"
description = "Prometheus exporter for Octopus Energy metrics. Works best when coupled with an Octopus Home Mini."
license = "AGPL-3.0-or-later"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gql==3.4.1",
    "httpx==0.25.1",
    "prometheus-client==0.22.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-jose==3.4.0",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
]
name = "octopus-usage-exporter"

[project.urls]
Repository = "https://github.com/josephrpalmer/octopus-usage-exporter.git"
Issues = "https://github.com/josephrpalmer/octopus-usage-exporter/issues"

[project.entry-points."octopus_usage_exporter.main"]
main = "app.octopus_usage_exporter:main"
